[MOV] website_slides: split models into files
Just to better see directory organization. This commit contains only
code move to try to keep history.

Move: channel tag and decorated m2m models (i.e. slide partner and
channel partner) models into their own file.

Task-4868193
