+++ addons/website_slides/models/__init__.py
@@ -2,12 +2,15 @@
 
 from . import gamification_challenge
 from . import gamification_karma_tracking
+from . import slide_slide_partner  # decorated m2m, import before
 from . import slide_slide
 from . import slide_question
 from . import slide_embed
+from . import slide_channel_partner  # decorated m2m, import before
 from . import slide_channel
 from . import slide_channel_tag
 from . import slide_slide_resource
+from . import slide_tag
 from . import res_config_settings
 from . import website
 from . import res_users
